Move relateManyProjectsToQfRound to another file to prevent build problem
